Rewrite to support multiple roku boxes.
This is a major re-write of the node server so that it can support
multiple Roku boxes.

The node layout has changed so that each box is a single node and
a command is used to launch applications on the box.

Both the node definition file and NLS file are dynamically
generated based on the applications installed on the roku boxes.